update shared and indexers
