+++ Composer/packages/lib/indexers/package.json
@@ -28,6 +28,7 @@
     "ts-jest": "^24.1.0"
   },
   "dependencies": {
+    "@bfc/shared": "*",
     "@bfcomposer/bf-lu": "1.1.2",
     "botbuilder-lg": "4.7.0-preview.93464",
     "botframework-expressions": "4.7.0-preview.93464",
