+++ Composer/packages/lib/indexers/src/index.ts
@@ -6,6 +6,5 @@ export * from './lgIndexer';
 export * from './luIndexer';
 export * from './type';
 export * from './diagnostic';
-export * from './lgUtils';
 export * from './dialogUtils';
 export * from './utils';
