+++ Composer/packages/lib/package.json
@@ -11,7 +11,7 @@
     "build:code-editor": "cd code-editor && yarn build",
     "build:shared": "cd shared && yarn build",
     "build:indexers": "cd indexers && yarn build",
-    "build:all": "yarn build:indexers && yarn build:shared && yarn build:code-editor"
+    "build:all": "yarn build:shared && yarn build:indexers && yarn build:code-editor"
   },
   "author": "",
   "license": "ISC"
