+++ Composer/packages/lib/shared/__tests__/lgUtils/parsers/parseLgTemplateName.test.ts
