+++ Composer/packages/lib/shared/package.json
@@ -31,7 +31,6 @@
     "@babel/plugin-transform-runtime": "^7.4.4",
     "@babel/preset-env": "^7.4.5",
     "@babel/preset-react": "^7.0.0",
-    "@bfc/indexers": "*",
     "@types/jest": "^24.0.11",
     "@types/nanoid": "^2.1.0",
     "@types/react": "16.9.0",
