+++ Composer/packages/lib/shared/src/index.ts
@@ -12,4 +12,4 @@ export * from './promptTabs';
 export * from './appschema';
 export * from './types';
 export * from './constant';
-export { LgMetaData, LgTemplateRef } from '@bfc/indexers';
+export * from './lgUtils';
