+++ Composer/packages/lib/shared/src/lgUtils/index.ts
@@ -4,3 +4,4 @@
 /** models */
 export { default as LgMetaData } from './models/LgMetaData';
 export { default as LgTemplateRef } from './models/LgTemplateRef';
+export { extractLgTemplateRefs } from './parsers/parseLgTemplateRef';
