+++ Composer/packages/lib/shared/src/lgUtils/models/LgTemplateRef.ts
