+++ Composer/packages/lib/shared/src/lgUtils/models/stringTypes.ts
