+++ Composer/packages/lib/shared/src/lgUtils/parsers/lgPatterns.ts
