+++ Composer/packages/lib/shared/src/lgUtils/parsers/parseLgTemplateName.ts
