+++ Composer/packages/lib/shared/src/lgUtils/stringBuilders/buildLgTemplateName.ts
