+++ Composer/packages/lib/shared/src/lgUtils/stringBuilders/buildLgTemplateRefString.ts
